chore: codebase cleanup
