+++ Makefile
@@ -8,7 +8,6 @@
 # 	endif
 # endif
 
-
 ARCH := UNSPECIFIED
 BUILD_DIR := bin
 SRC_DIR := src
@@ -36,14 +35,12 @@ kernel:
 	mkdir -p $(OBJ_DIR_RES)
 	$(ASM) $(KERNEL_SRC_DIR)/idt.asm -o $(OBJ_DIR)/idt.o
 	$(ASM) $(KERNEL_SRC_DIR)/gdt.asm -o $(OBJ_DIR)/gdt.o
-	$(ASM) $(KERNEL_SRC_DIR)/interrupts.asm -o $(OBJ_DIR)/ISR.o
-	$(ASM) $(KERNEL_SRC_DIR)/memset.asm -o $(OBJ_DIR)/MemSet.o
-	$(ASM) $(KERNEL_SRC_DIR)/paging.asm -o $(OBJ_DIR)/paging.o
-	$(ASM) $(KERNEL_SRC_DIR)/pushpop.asm -o $(OBJ_DIR)/pushpop.o
+	$(ASM) $(KERNEL_SRC_DIR)/interrupts.asm -o $(OBJ_DIR)/interrupts.o
+	$(ASM) $(KERNEL_SRC_DIR)/memory.asm -o $(OBJ_DIR)/memory.o
 	$(ASM) $(KERNEL_SRC_DIR)/kernel_entry.asm -o $(OBJ_DIR)/kernel_entry.o
 	$(CXX) -MF$(OBJ_DIR)/idt.cxx.o.d -o $(OBJ_DIR)/idt.cxx.o -c $(KERNEL_SRC_DIR)/idt.cxx
 	$(CXX) -MF$(OBJ_DIR)/pic.cxx.o.d -o $(OBJ_DIR)/pic.cxx.o -c $(KERNEL_SRC_DIR)/pic.cxx
-	$(CXX) -MF$(OBJ_DIR)/paging.cxx.o.d -o $(OBJ_DIR)/paging.cxx.o -c $(KERNEL_SRC_DIR)/paging.cxx
+	$(CXX) -MF$(OBJ_DIR)/memory.cxx.o.d -o $(OBJ_DIR)/memory.cxx.o -c $(KERNEL_SRC_DIR)/memory.cxx
 	$(CXX) -MF$(OBJ_DIR)/io.cxx.o.d -o $(OBJ_DIR)/io.cxx.o -c $(KERNEL_SRC_DIR)/io.cxx
 	$(CXX) -MF$(OBJ_DIR)/terminal.cxx.o.d -o $(OBJ_DIR)/terminal.cxx.o -c $(KERNEL_SRC_DIR)/terminal.cxx
 	$(CXX) -MF$(OBJ_DIR)/debugging.cxx.o.d -o $(OBJ_DIR)/debugging.cxx.o -c $(KERNEL_SRC_DIR)/debugging.cxx
@@ -55,10 +52,8 @@ kernel:
 	$(CXX) -MF$(OBJ_DIR)/kmain.cxx.o.d -o $(OBJ_DIR)/kmain.cxx.o -c $(KERNEL_SRC_DIR)/kmain.cxx -DARCH=\"$(ARCH)\"
 	$(CXX) -MF$(OBJ_DIR)/bootinfo.cxx.o.d -o $(OBJ_DIR)/bootinfo.cxx.o -c $(KERNEL_SRC_DIR)/bootinfo.cxx
 	$(CXX) -MF$(OBJ_DIR)/power.cxx.o.d -o $(OBJ_DIR)/power.cxx.o -c $(KERNEL_SRC_DIR)/power.cxx
-	$(OBJCOPY) $(FONTS_DIR)/font.psf $(OBJ_DIR_RES)/font.o
-	$(CXX_LINK) -o $(BUILD_DIR)/microCORE.kernel $(OBJ_DIR)/kernel_entry.o $(OBJ_DIR)/kmain.cxx.o $(OBJ_DIR)/bootinfo.cxx.o $(OBJ_DIR)/power.cxx.o $(OBJ_DIR)/pushpop.o $(OBJ_DIR)/gdt.o $(OBJ_DIR)/idt.o $(OBJ_DIR)/ISR.o $(OBJ_DIR)/MemSet.o $(OBJ_DIR)/paging.o $(OBJ_DIR)/idt.cxx.o $(OBJ_DIR)/io.cxx.o $(OBJ_DIR)/paging.cxx.o $(OBJ_DIR)/terminal.cxx.o $(OBJ_DIR)/pic.cxx.o $(OBJ_DIR)/debugging.cxx.o $(OBJ_DIR)/kutil.cxx.o $(OBJ_DIR)/kbd.cxx.o $(OBJ_DIR_RES)/font.o $(OBJ_DIR)/gfx.cxx.o $(OBJ_DIR)/string.cxx.o $(OBJ_DIR)/uart.cxx.o -T $(RES_DIR)/Linkerscript
+	$(CXX_LINK) -o $(BUILD_DIR)/microCORE.kernel $(OBJ_DIR)/kernel_entry.o $(OBJ_DIR)/kmain.cxx.o $(OBJ_DIR)/bootinfo.cxx.o $(OBJ_DIR)/power.cxx.o $(OBJ_DIR)/gdt.o $(OBJ_DIR)/idt.o $(OBJ_DIR)/interrupts.o $(OBJ_DIR)/memory.o $(OBJ_DIR)/idt.cxx.o $(OBJ_DIR)/io.cxx.o $(OBJ_DIR)/memory.cxx.o $(OBJ_DIR)/terminal.cxx.o $(OBJ_DIR)/pic.cxx.o $(OBJ_DIR)/debugging.cxx.o $(OBJ_DIR)/kutil.cxx.o $(OBJ_DIR)/kbd.cxx.o $(OBJ_DIR_RES)/font.o $(OBJ_DIR)/gfx.cxx.o $(OBJ_DIR)/string.cxx.o $(OBJ_DIR)/uart.cxx.o -T $(RES_DIR)/Linkerscript
 
-	
 clean:
 	rm -rfv $(BUILD_DIR) iso
