+++ include/kernel/gdt.hxx
@@ -1,0 +1,5 @@
+//
+// Created by rizet on 1/26/21.
+//
+#pragma once
+extern "C" void loadGDT64();
