--- include/kernel/mem.hxx
@@ -1,4 +1,0 @@
-#pragma once
-
-void beginPaging();
-
